Add exchange functionality to transfers

+++ src/main/java/org/kaczucha/account/application/port/AccountServiceUseCase.java
@@ -3,16 +3,18 @@ package org.kaczucha.account.application.port;
 import org.kaczucha.account.web.dto.AccountRequest;
 import org.kaczucha.account.web.dto.AccountResponse;
 
+import java.math.BigDecimal;
+
 public interface AccountServiceUseCase {
     void save(AccountRequest accountRequest);
 
     AccountResponse findById(long id);
 
-    void transfer(long fromId, long toId, double amount, String currency);
+    void transfer(long fromId, long toId, BigDecimal amount, String currency);
 
-    void validateAmount(double amount);
+    void validateAmount(BigDecimal amount);
 
     void deleteById(Long id);
 
-    void withdraw(long id, double amount);
+    void withdraw(long id, BigDecimal amount);
 }
